Added vim autosetup.

+++ install.sh
@@ -1,19 +1,18 @@
 #!/bin/bash
 DOTFILES_DIR=${DOTFILES_DIR:-src/dotfiles}
-FILES=${FILES:-bashrc profile profile.d gitconfig global-gitignore vim vimrc}
+FILES=${FILES:-bashrc profile profile.d/google.sh gitconfig global-gitignore vim vimrc}
+PROFILE_D=${PROFILE_D:-.profile.d}
 backup() {
     local file=$1
-    if ! [ -L .$file ]; then
+    if ! [ -L .$file ] && [ -f .$file ]; then
         mv .$file .$file.orig
-        true
     fi
-    false
 }
 
 link() {
     local file=$1
 
-    ln -sf $DOTFILES_DIR/$file .$file
+    [ -L .$file ] || ln -sf $DOTFILES_DIR/$file .$file
 }
 
 vim_setup() {
@@ -25,14 +24,17 @@ vim_vundle() {
     vim +PluginInstall +qall
 }
 
+[ -d ~/${PROFILE_D} ] || mkdir ~/${PROFILE_D}
 echo -n "Creating/Updating links to dotfiles"
 for file in $FILES; do
-    backup $file && \
+    backup $file
     link $file
     echo -n "."
 done
 echo "(OK)"
 
-echo -n "Application specific stuff "
-vim_setup
-echo "(OK)"
+if [ -z "$NO_VIM" ]; then
+    echo -n "Application specific stuff "
+    vim_setup
+    echo "(OK)"
+fi
